Merges views.py and reservationbook.py, integrates all functionality together (with exception of repeated reservation and updating from waiting -> reserved), refactors a ton of things, reviews all unit tests, some commented out until integration is complete.
