+++ app/constants.py
@@ -1,0 +1,7 @@
+STATUS_CODE = {
+    'OK': 200,
+    'UNAUTHORIZED': 401,
+    'NOT_FOUND': 404,
+    'UNPROCESSABLE': 422
+}
+MAX_EQUIPMENT_NUMBER = 2
