+++ app/core/equipment.py
@@ -23,3 +23,30 @@ class Equipment:
 
 	def setId(self, equipmentId):
 		self.equipmentId = equipmentId
+
+	def getNumLaptops(self):
+		return self.equipment['laptops']
+
+	def setNumLaptops(self, laptops):
+		self.equipment['laptops'] = laptops
+
+	def getNumProjectors(self):
+		return self.equipment['projectors']
+
+	def setNumProjectors(self, projectors):
+		self.equipment['projectors'] = projectors
+
+	def getNumWhiteboards(self):
+		return self.equipment['whiteboards']
+
+	def setNumWhiteboards(self, whiteboards):
+		self.equipment['whiteboards'] = whiteboards
+
+	def to_dict(self):
+		equipment_data = {}
+		equipment_data['equipmentId'] = self.getId()
+		equipment_data['laptops'] = self.equipment['laptops']
+		equipment_data['projectors'] = self.equipment['projectors']
+		equipment_data['whiteboards'] = self.equipment['whiteboards']
+		return equipment_data
+
