+++ app/views.py
@@ -3,21 +3,14 @@ from .decorators import *
 from flask import request, jsonify
 from app.mapper import ReservationMapper
 from app.mapper import WaitingMapper
-from app.mapper import RoomMapper
-from app.mapper import UserMapper
 from app.mapper import TimeslotMapper
 from app.mapper import EquipmentMapper
+from app.mapper import UserMapper
+from app.reservationbook import ReservationBook
 from datetime import datetime
-import calendar
-from uuid import uuid4
+from constants import STATUS_CODE
 
-STATUS_CODE = {
-    'OK': 200,
-    'UNAUTHORIZED': 401,
-    'NOT_FOUND': 404,
-    'UNPROCESSABLE': 422
-}
-
+reservationbook = ReservationBook()
 
 ##########
 # ROUTES #
@@ -38,7 +31,33 @@ def internal_error(error):
 def login():
     if request.method == 'POST':
         data = request.get_json();
-        return validate_login(data)
+        if 'username' not in data or 'password' not in data:
+            response = jsonify({'login error': '`username` and `password` fields are required.'})
+            response.status_code = STATUS_CODE['UNPROCESSABLE']
+            return response
+
+        user = UserMapper.find(str(data['username']))
+        if not user:
+            response = jsonify({'login error': 'That user does not exist.'})
+            response.status_code = STATUS_CODE['NOT_FOUND']
+            return response
+
+        if user.getPassword() != str(data['password']):
+            response = jsonify({'login error': 'Credentials refused for that user.'})
+            response.status_code = STATUS_CODE['UNAUTHORIZED']
+            return response
+
+        session['logged_in'] = True
+        session['username'] = user.getId()
+
+        success = {
+            'login success': 'Successfully logged in',
+            'data': {
+                'username': str(user.getId()),
+                'capstone': str(user.isCapstone())
+            }
+        }
+        return jsonify(success)
 
     if request.method == 'GET':
         return is_logged_in()
@@ -56,13 +75,8 @@ def logout():
 @require_login
 def get_all_rooms():
     if request.method == 'GET':
-        room_models = RoomMapper.findAll()
-        if room_models:
-            rooms = sorted([room.getId() for room in room_models])
-        else:
-            rooms = []
         roomdata = {
-            'rooms': rooms
+            'rooms': reservationbook.get_all_rooms()
         }
         return jsonify(roomdata)
 
@@ -73,7 +87,10 @@ def get_all_rooms():
 def make_new_reservation():
     if request.method == 'POST':
         data = request.get_json()
-        return validate_new_reservation(data)
+        response = validate_reservation_payload_format(data)
+        if response:
+            return response
+        return reservationbook.make_new_reservation(data)
 
 
 @app.route('/reservations/room/<roomId>', methods=['GET'])
@@ -81,22 +98,7 @@ def make_new_reservation():
 @require_login
 def get_reservations_by_room(roomId):
     if request.method == 'GET':
-        reservations = ReservationMapper.findByRoom(roomId)
-        reservations_data = []
-        if reservations:
-            for reservation in reservations:
-                reservations_data += [reservation.to_dict()]
-        data = {
-            'roomId': roomId,
-            'reservations': reservations_data
-        }
-        waitings = WaitingMapper.findByRoom(roomId)
-        waitings_data = []
-        if waitings:
-            for waiting in waitings:
-                waitings_data += [waiting.to_dict()]
-        data.update({'waitings': waitings_data})
-        return jsonify(data)
+        return reservationbook.get_reservations_by_room(roomId)
 
 
 @app.route('/reservations/user/<username>', methods=['GET'])
@@ -104,122 +106,26 @@ def get_reservations_by_room(roomId):
 @require_login
 def get_reservations_by_user(username):
     if request.method == 'GET':
-        reservations = ReservationMapper.findByUser(str(username))
-        reservations_data = []
-        if reservations:
-            for reservation in reservations:
-                reservations_data += [reservation.to_dict()]
-        data = {
-            'username': username,
-            'reservations': reservations_data
-        }
-        waitings = WaitingMapper.findByUser(str(username))
-        waitings_data = []
-        if waitings:
-            for waiting in waitings:
-                waitings_data += [waiting.to_dict()]
-        data.update({'waitings': waitings_data})
-        return jsonify(data)
-
+        return reservationbook.get_reservations_by_user(username)
 
 @app.route('/reservations/all', methods=['GET'])
 @nocache
 @require_login
 def get_all_reservations():
     if request.method == 'GET':
-        reservations = ReservationMapper.findAll()
-        reservations_data = []
-        if reservations:
-            for reservation in reservations:
-                reservations_data += [reservation.to_dict()]
-        data = {
-            'reservations': reservations_data
-        }
-        waitings = WaitingMapper.findAll()
-        waitings_data = []
-        if waitings:
-            for waiting in waitings:
-                waitings_data += [waiting.to_dict()]
-        data.update({'waitings': waitings_data})
-        return jsonify(data)
-
+        return reservationbook.get_all_reservations()
 
 @app.route('/reservations/<reservationId>', methods=['DELETE'])
 @nocache
 @require_login
 def delete_reservation(reservationId):
     if request.method == 'DELETE':
-        reservation = ReservationMapper.find(reservationId)
-        waiting = WaitingMapper.find(reservationId)
-        if not reservation and not waiting:
-            response = jsonify({'reservation error': 'that reservationId does not exist'})
-            response.status_code = STATUS_CODE['NOT_FOUND']
-            return response
-
-        if reservation:
-            timeslotId = reservation.getTimeslot().getId()
-            equipmentId = reservation.getEquipment().getId()
-            ReservationMapper.delete(reservationId)
-            ReservationMapper.done()
-            TimeslotMapper.delete(timeslotId)
-            TimeslotMapper.done()
-            EquipmentMapper.delete(equipmentId)
-            EquipmentMapper.done()
-            data = {
-                'success': 'reservation successfully deleted',
-                'reservationId': reservationId
-            }
-        if waiting:
-            timeslotId = waiting.getTimeslot().getId()
-            equipmentId = waiting.getEquipment().getId()
-            WaitingMapper.delete(reservationId)
-            WaitingMapper.done()
-            TimeslotMapper.delete(timeslotId)
-            TimeslotMapper.done()
-            EquipmentMapper.delete(equipmentId)
-            EquipmentMapper.done()
-            data = {
-                'success': 'reservation on waiting list successfully deleted',
-                'waitingId': reservationId
-            }
-
-        # update reservation and waiting lists here
-        return jsonify(data)
+        return reservationbook.delete_reservation(reservationId)
 
 
 ####################
 # HELPER FUNCTIONS #
 ####################
-
-def validate_login(data):
-    if 'username' not in data or 'password' not in data:
-        response = jsonify({'login error': '`username` and `password` fields are required.'})
-        response.status_code = STATUS_CODE['UNPROCESSABLE']
-        return response
-
-    user = UserMapper.find(str(data['username']))
-    if not user:
-        response = jsonify({'login error': 'That user does not exist.'})
-        response.status_code = STATUS_CODE['NOT_FOUND']
-        return response
-
-    if user.getPassword() != str(data['password']):
-        response = jsonify({'login error': 'Credentials refused for that user.'})
-        response.status_code = STATUS_CODE['UNAUTHORIZED']
-        return response
-
-    session['logged_in'] = True
-    session['username'] = user.getId()
-
-    success = {
-        'login success': 'Successfully logged in',
-        'data': {
-            'username': str(user.getId()),
-            'capstone': str(user.isCapstone())
-        }
-    }
-    return jsonify(success)
-
 
 def is_logged_in_bool():
     return 'logged_in' in session and 'username' in session and session['logged_in']
@@ -240,68 +146,6 @@ def unauthorized():
     response = jsonify({'unauthorized': 'Not logged in. You must login.'})
     response.status_code = STATUS_CODE['UNAUTHORIZED']
     return response
-
-
-def validate_new_reservation(data):
-    response = validate_reservation_payload_format(data)
-    if response:
-        return response
-
-    startTime = int(data['timeslot']['startTime'])
-    endTime = int(data['timeslot']['endTime'])
-    date = str(data['timeslot']['date'])
-    dateList = date.split('/')
-    roomId = data['roomId']
-    username = data['username']
-    description = str(data['description'])
-    laptop = int(data['equipment']['laptop'])
-    board = int(data['equipment']['board'])
-    projector = str(data['equipment']['projector'])
-
-
-    response = validate_make_new_reservation_date(dateList)
-    if response:
-        return response
-
-    response = validate_reservation_room_user_exists(roomId, username)
-    if response:
-        return response
-
-    reservations = ReservationMapper.findAll()
-
-    # no use for `block` parameter, for now, just passing empty strin
-    time = TimeslotMapper.makeNew(startTime, endTime, datetime(int(dateList[0]), int(dateList[1]), int(dateList[2])), '', username, str(uuid4()))
-    TimeslotMapper.done()
-    room = RoomMapper.find(roomId)
-    user = UserMapper.find(username)
-    equipment = EquipmentMapper.makeNew(laptop, projector, board, str(uuid4()))
-    EquipmentMapper.done()
-
-
-    if validate_make_new_reservation_timeslots(reservations, dateList, startTime, endTime):
-        return jsonify(commit_new_waiting(room, user, time, description, equipment))
-    else:
-        return jsonify(commit_new_reservation(room, user, time, description, equipment))
-
-
-def commit_new_waiting(room, user, time, description, equipment):
-    waiting = WaitingMapper.makeNew(room, user, time, description, equipment, str(uuid4()))
-    WaitingMapper.done()
-    response_data = {
-        'makeNewReservation': 'there is a conflict: added to the waitlist',
-        'waitingId': waiting.getId()
-    }
-    return response_data
-
-
-def commit_new_reservation(room, user, time, description, equipment):
-    reservation = ReservationMapper.makeNew(room, user, time, description, equipment, str(uuid4()))
-    ReservationMapper.done()
-    response_data = {
-        'makeNewReservation': 'successfully created reservation',
-        'reservation': reservation.getId()
-    }
-    return response_data
 
 
 def validate_reservation_payload_format(data):
@@ -359,8 +203,9 @@ def validate_reservation_payload_format(data):
         response.status_code = STATUS_CODE['UNPROCESSABLE']
         return response
 
+    date = str(data['timeslot']['date'])
+    dateList = date.split('/')
 
-def validate_make_new_reservation_date(dateList):
     def date_error():
         response = jsonify({'makeNewReservation error': '`date` must be in the format 2034/03/03 and must be a valid date, today or in the future.'})
         response.status_code = STATUS_CODE['UNPROCESSABLE']
@@ -382,31 +227,3 @@ def validate_make_new_reservation_date(dateList):
     current_date = datetime(now.year, now.month, now.day)
     if payload_date < current_date:
         return date_error()
-
-
-def validate_reservation_room_user_exists(roomId, username):
-    if not UserMapper.find(username) or not RoomMapper.find(roomId):
-        response = jsonify({'makeNewReservation error': 'Either the room or user does not exist.'})
-        response.status_code = STATUS_CODE['NOT_FOUND']
-        return response
-
-
-def validate_make_new_reservation_timeslots(reservations, dateList, startTime, endTime):
-    def to_timestamp(date_list, time):
-        return calendar.timegm(datetime(int(date_list[0]), int(date_list[1]), int(date_list[2]), int(time)).timetuple())
-
-    new_timestamp_start = to_timestamp(dateList, startTime)
-    new_timestamp_end = to_timestamp(dateList, endTime)
-
-    if reservations:
-        for reservation in reservations:
-            timeslot = reservation.getTimeslot()
-            timeslot_date_list = timeslot.getDate().strftime('%Y/%m/%d').split('/')
-
-            existing_timestamp_start = to_timestamp(timeslot_date_list, timeslot.getStartTime())
-            existing_timestamp_end = to_timestamp(timeslot_date_list, timeslot.getEndTime())
-
-            if (new_timestamp_start < existing_timestamp_end) and \
-                    (new_timestamp_end > existing_timestamp_start):
-                return True
-    return False
